requirements.txt
Django>=1.5
python-mpd2
requests
djangorestframework
